installer: fix parsing gcr token as git repo parameter
Previously git parameter was specified before gcr token parameter. This
resulted in a behaviour that if we did not specify GCR token and just
specified git repo, git repo name was considered as  GCR tokken by the
script, thus token was incorrect, and git was not specified.
Solution is to put all optional parameters at the end.